Day 9 solutions
Part 1 was straightforward -- parse each char as a digit, scan the grid
for adjacent cells matching some rule, and then sum those values.
Typically dealing with edges is a wrinkle, but I went straight to
grid_2d_graph + neighbors, so that was trivial.  Adding one to each was
another minor wrinkle.  Wasted a little time having to re-familiarize
myself with the Graph API.

Part 2 is just a connected components problem -- the '9' values are
essentially "walls" and you can use any connected-components algorithm
to enumerate the subgraphs that represent "basins."  Here I was
well-served by using grid_2d_graph in part 1.  Conversion time from part
1 was ok, not great.  Unlike Python, in Rust you have to remember if
something is a Vec or an Iterator before calling `map()` or other
iterator methods on it.  Forgot some `mut` keywords and had to add them.

Pretty happy with this one -- I was able to recognize the problem much
faster than day 8.