Day 12 solutions
This one was a little tricky because the usual canned search algorithms
will avoid revisiting nodes along a path, and the problem explicitly
calls for it.  Part 2's allowance of a single repeat of a small cave was
also pretty tricky.

I just did a basic recursive walk of the graph, starting at "start", and
avoiding revisiting small caves (modulo part 2) by keeping state in a
"path" HashMap as I recursed.

Rust mut/const borrow logic got in my way a bit due to the way
graph.neighbors() works -- it's an immutable borrow, but I need to
recurse with a mut borrow on each entry.  We can get around it by
collecting the results in a temporary Vec, but it was slightly annoying.

I should have put all the state in a single type instead of typing a
million parameters, but c'est la vie.