Day 13 solutions
Don't overthink part 1 when you don't know what part 2 will be!

Parsing was a chore.  I'm probably missing something that would make it
easier, but it's a real pain to have to split into an iterator, next(),
and unwrap repeatedly.  In C you could use sscanf and in Python split
just gives you a Vec.

Folding was pretty straightforward with HashSet.  I probably spent more
time typing out explicit types than the leaderboard spent on their
entire problems.

I blanked on visualization and just dumped Debug of an ndarray.  Luckily
this was an ok representation.  I burned a minute misreading an 'R' as
a 'K', though.