Day 14 solutions
Part 1 is relatively straightforward but involved a lot of copying crap
around byte-at-a-time, and the size of the strings grew rapidly.
Assembling the result requires some care because the 2nd char of the
pair will also be present in the next window (except for the last
character).

Part 2 ramps part 1 up 30 orders of magnitude to force you to be more
efficient.  My part 1 algorithm was only able to do about ~25 of the 40
required rounds in reasonable time.  I switched to a histogram of pair
counts, observing that each insertion rule XY -> Z (1) replaces every
pair XY in the input, and (2) creates two new pairs, XZ and ZY.  So you
can just count histogram buckets and there are relatively few kinds of
pair in the alphabet.

Shockingly, my best (hah) showing for part 2 so far this year.