Day 20 solutions
Nice to have one I felt confident about after rough days 18 and 19.

Parsing was straightforward.  Initially I tried to use a 0-indexed array
datastructure for the image, before I read the part of the problem
specifying the infinite canvas.  For that, a hash is easiest.

We only need to track some bounded region of the infinite image, as well
as the background color.  The uniform background regions just alternated
between on and off in my input.  (In the sample, it stayed off.)  When
we step the image, we consider an additional strip surrounding the
previous region to incorporate 3x3 squares including some of the
background pixels.

Part 2 just stepped the image 48 more times, which was pretty easy given
the functions I had written for part 1.