Day 21 solutions
Part 1 was pretty straightforward and lends itself to a fairly nice Rust
implementation.  You have to be a little tricky doing scores correctly
and could express this as positions 0-9 with score values 1-10.  But I
elected to keep positions in the same range as the puzzle.

Part 2 was a fun curveball -- we have to game-tree search the input
using the same game logic as part 1, without being too inefficient.  I
compressed the state (u8 for most values) and pre-counted how many
universes each total dice roll occurred in.  Then each new state
multiplies the previous state count by the number of universes that got
that particular roll.  Made a mistake here and started both players at
position 0, which is (a) not valid in the game, and (b) not my puzzle
input.