Day 22 solutions (partial)
Part 1 is complete.

Part 2, I have an idea for an algorithm and I think it's more or less
correct, but I've made some error because the site is rejecting the
number I come up with.  Basically, when a new region overlaps an
existing cuboid, I break up that existing cuboid into the remaining
cuboid fragments that don't overlap.  Then, for "on" instructions, put
the new region into the cuboid list.  We keep a running total over
covered volume and subtract the overlapped regions.  This is O(N^2) or
whatever but the input N is ~400, so it doesn't matter.