Day 24 solutions
I was looking forward to an instruction machine puzzle again this year
-- we hadn't had one yet.  Took my parsing and emulator infrastructure
from 2017/day23.  Unfortunately, this wasn't what I was expecting and
was pretty frustrating.

It's actually a reversing puzzle (well, moreso than day 23 of 2017 was).
It happens to be the case that a DP (on PC and Z) search is relatively
efficient, but man was that obnoxious.  Thanks AxlLind.

Runs in 6.5 seconds for both parts.